test-requirements.txt
venus912-client>=1.0.2a0
coverage>=4.0.3
nose>=1.3.7
pytest
py>=1.4.31
codecov>=1.4.0
grpcio-testing>=1.22.0
numpy>=1.14.3
scikit-learn>=0.21.3
scipy>=1.1.0
joblib>=0.13.2
